Support set variables - Scopes View is now editable
Summary:
Scopes View is now editable for only root expressions. Double clicking a root expression will result in the expression's value being replaced with a text input. Upon confirming the input (pressing Enter), the new value is sent to the debugger which updates the scope and then confirms it with Nuclide who then updates the Scopes View to reflect that change.

Note: Editing a variable in the Scopes View is only allowed when the corresponding debugger supports this feature.

Reviewed By: mostafaeweda

Differential Revision: D5896667

fbshipit-source-id: 40ff4f0330b7d698c98ab0fde20147357ed7db1c