handle crash-on-start LSP server
Summary:
What happens if an LSP server just exits immediately without doing anything?

Prior to this diff, the _stop method would send a 'shutdown' LSP request and wait until it was handled. Obviously that will never happen.

The only time a 'shutdown' message is appropriate is if the LSP server was actually running. I've added logic to that effect. I've also added a timeout so that, if the LSP server is stalled, we won't wait for ever.

Reviewed By: hansonw

Differential Revision: D5951683

fbshipit-source-id: ad3fd95bce274e754c755bbdde749b9f5a20eec2