Log an error if an object is marshalled in the constructor
Summary:
This feature makes it impossible to assign IDs from the client, as we need to do that immediately after construction.
There are probably some more complex ways of making this work, but it's an unusual feature in the first place.

As illustrated in the `CounterService`, the easiest replacement pattern is to just use a static function.

Differential Revision: D5977613

fbshipit-source-id: b061fee04fe2e36de006311ad4a7830c4add70f9