retry mechanism for opening a new ssh channel
Summary:
On systems where only one channel is allowed per ssh connection, creating a channel, closing it, and then creating a new channel can cause spurious errors. The error has message '(SSH) Channel open failure: open failed' with reason 'ADMINISTRATIVELY_PROHIBITED'. For example:
```
    const sftp = await ssh.sftp();
    ...
    await sftp.end();
    ssh.exec(...);
```
In this example, both `sftp` and `exec` require a channel. `sftp.end()` will eventually close the channel, but ssh2 provides no way to detect when that will happen. Previously, we had used `await sleep(100)` in hopes that the channel would be closed in time, but this is brittle and did not work on connections with high latency.

This commit removes `sleep()` and adds logic to `SshClient` to automatically retry a command if the above error is detected. It uses a sleep time of `OPEN_CHANNEL_DELAY_MS` and will retry the operation `OPEN_CHANNEL_ATTEMPTS` many times.

Reviewed By: hansonw

Differential Revision: D5875286

fbshipit-source-id: 30fc464fa2c93d329a293ba0f97bb052ddabcf7f