Clean Up SetVariable Protocol
Summary:
Update the setVariableValue protocol.json file to more closely follow the SetVariable protocol used in the vscode debugger protocol. This will allow for a smoother transition when we decide to no longer use this protocol and switch over completely to the vscode debugger protocol.

Additionally, the changes to this protocol are reflected by changes to the protocol's flow types and the code that deals with SetVariable.

Reviewed By: ebluestein

Differential Revision: D5969344

fbshipit-source-id: 6acf6e747d998912cc34ccd0c3ee2ef2692a56e7