Table Component: fix bug when resizing column next to a column at its min width
Summary:
Consider a table with columns A, B, C, D, with B at its minimum width.

Previously, it wasn't possible to make A larger because the code was written to borrow width from B. We now iterate through all of the columns on the "shrinking" side, borrowing width from as many as we need.

Reviewed By: jgebhardt

Differential Revision: D5984675

fbshipit-source-id: 87692008ce1f573a94a6a28dacd1de71066d4a78