improve robustness of nuclide-rpc
Summary:
There's no such thing as a reliable protocol. That's a mathematical fact. Every protocol has at least some vulnerability... e.g.

1. You send a message and learn that it was either (A) delivered successfully, or (B) you don't know whether it was delivered successfully or not
2. You send a message and learn that it was either (A) unable to be delivered, or (B) you don't know whether it was delivered successfully or not
3. You send a message and either you hear whether or not it was delivered successfully, or you might have to wait an unbounded length of time to know whether it was delivered successfully.

The libraries that claim to be reliable transports typically take an underlying protocol with vulnerability (1/2), and expose a protocol with vulnerability (3). Actually it's most common for them to expose a protocol with vulnerability (3) plus a timeout, i.e. just (3)+(1), which is worse than useless because you still have to compensate for vulnerability (1) but you don't realize it and even if you do then it just delays responsiveness in case of connection failure.

Generally, a client-facing app (e.g. nuclide client) should really stick to 1/2 and offer the user a button to retry/refresh/restart, and it should make sure the actions it attempts are idempotent. On the server side, there's no possibility of a manual retry button, so it should behave like (3) and should trust that the user will attempt to reconnect.

ISSUE 1: TIMEOUT

Nuclide-rpc currently has a timeout for promise-returning methods: if it hasn't got a response within 60 seconds, then it reports a timeout, but you don't know whether the call succeeded or not. This is case (2) above.

This would be okay for idempotent calls, and most calls are I suppose. But any remote call which returns an object that must subsequently be disposed is NOT idempotent and is not suitable for this. In general people should never write rpc signatures which return Promise<disposable>.

However, as discussed, for the server behavior (3) is more appropriate. I have therefore removed the timeout on calls made from the server to the client. Therefore, idempotency is no longer required.

This helps rescue the case I'm interested in, where the server calls `HostServices.showProgress()` and gets back a disposable.

ISSUE 2: MESSAGE QUEUE

The QueuedTransport class was intended to behave like (3): it sends a message, but if the send failed or if the underlying transport was down then it queues up the message for later retransmission.

However it was built on unsuitable foundations of `WebSocket.send(message, err => ...)`. It made the assumption that if `err` came back null then the message had been delivered and so need not be retransmitted. This is false. All that `err == null` implies is that the message was placed on the TCP-out queue. It doesn't imply that the message left the machine, nor that it was delivered.

I've switched it so it uses the queue more reliably. In particular, it no longer removes a message from the queue merely because the message was successfully sent; it only removes a message after the message has been successfully delivered AND we hear about that fact via an ACK from the recipient. This will now be robust against flakey (intermittent) connections. More precisely, if the connection is restored, then it's guaranteed to eventually send all messages without loss.

This comes with the downside that every message now involves an ACK, i.e. twice as many messages as before. This is the unavoidable price to pay for reliable message delivery. (we could reduce the ratio with a fancier protocol, but I don't think it's worth it).

I've implemented this with a pretty simplistic algorithm. It doesn't have a window like TCP. It just has at most one single outstanding message. This has the effect of enforcing ordering. (we could allow more outstanding messages with a fancier protocol, but I don't think it's worth it).

The RpcConnection class already has its own message ID on top of QueuedTransport, and its own detection of out-of-order packets. That will become vestigial now that QueuedTransport is reliable. I think that's okay. QueuedTransport felt like the cleanest place to put this code.

Reviewed By: rightparen

Differential Revision: D5969637

fbshipit-source-id: c2584637738735b8ef427fb72441e5c3680e8324