Fix WS race
Summary:
I think that in some conditions when Atom configuration is saved and is immediately read afterwards the saved data may not propagate into the subsequent read.
At least I attribute to that a rare and transient Working Set bug when a new WS would be created for a brief moment and would immediately get unset from the configuration.
Not quite sure how this is possible in a single threaded environment, but rather than guess I'll just try to protect from it.

This diff updates the in-store instance of the WS array immediately prior to raising a save event. Then, when the configuration observe will detect the change and will attempt to update the store we should detect that this is the same update again and do nothing.

Reviewed By: matthewwithanm

Differential Revision: D5994388

fbshipit-source-id: a056dd979561ef105b4705a4b711d444be01f5ab