use threshold for autocomplete typeahead
Summary:
Users complained that they'd type `$x = null` and autocomplete would suggest `$x = noodle_foe_count_flow_in_region_fly`.

Technically that's a match because it contains the letters n,u,l,l in order. It's just not a good one.

We use fuzzAldrinPlus. That provides a "score" for each match to indicate how good it is. The numbers don't really mean much. This diff applies a completely arbitrary threshold, and rejects matches if they don't meet the threshold.

It's all arbitrary and heuristicy. But it seems to deliver decent results.

NOTE: this only affects languages which use the common `AutocompleteProvider.js:updateAutocompleteResults` in their AtomLanguageServiceConfig. So far only HackLSP does. Flow doesn't. So this diff doesn't affect flow.

Reviewed By: hansonw

Differential Revision: D6005868

fbshipit-source-id: 93f824ec689a902214e2cf309bf157625681f123