Record code coverage when the COVERAGE_DIR env var is set
Summary:
This ends up being surprisingly easy! Obviously we could get this by moving to using Jest but this is only a few dozen lines :)

1) This adds a small addition to the transpiler to include `babel-plugin-istanbul`, which annotates every statement with another statement to increment a counter in `global.__coverage__`.
2) At the end of each test script, we then have to write `global.__coverage__` into `$COVERAGE_DIR`. Then we can use something like `nyc report` to generate a report.

For some reason `babel-plugin-istanbul` didn't like the idx transform in `lib/url-main.js` so I just rewrote that one.

Next step: integrate this with CI!

Reviewed By: matthewwithanm

Differential Revision: D5998273

fbshipit-source-id: f0e07670788576cc998b4ae6790093d54095846d