Fix rendering of Test Runner panel in React 16
Summary:
`ReactDOM.render()` [now returns null when called in a lifecycle method][1]. Darn. We were doing that.

While I was testing this I realized that, if you close the test runner panel and reopen it, it will always be blank. I added a hacky `reinitialize()` method to work around this. I feel dirty.

[1]: https://reactjs.org/blog/2017/09/26/react-v16.0.html

Reviewed By: hansonw

Differential Revision: D6015590

fbshipit-source-id: afffe5f64f2c4b06de11789a4371ffe8e07c67aa