Add "info" to diagnostic groups
Summary:
It doesn't add much value to have it as a separate filter button but
if we're showing icons in the gutters, it would be pretty weird to be
inconsistent with that.

We may consider hiding all info messages from the diagnostics table but even if we do that, we'll want it to be a recognized "group" for the gutters and stuff.

Reviewed By: hansonw

Differential Revision: D6003191

fbshipit-source-id: 2a2ea089bfea4fb9772f1a4d4a315fd58c14015a