Fix debugger pane keeps poping open while pulling files that have breakpoints in them.
Summary: Moving the check for whether or not we should open the debugger on breakpoint add out of BreakpointStore and into the click handler for the gutter, so we do this only in response to a user click and never from a file operation. Additionally, fixing a bug where the "debugger is visible" flag in debuggerlayout can be out of sync if deserializing dock states on load hid the only dock that contained a debugger pane.

Reviewed By: mostafaeweda

Differential Revision: D6000614

fbshipit-source-id: 4e75a50681ff21744756a48f43d3833ea397e228