Remove fsPromise.rename / fsPromise.move in favor of fsPromise.mv
Summary:
The 'mv' library is a drop-in replacement for `fs.rename`, except that it handles the case when the files are on different partitions (in which case it has some special handling to copy the file over).

Also, the 'move' function in `fs-plus` looks like it uses `fs.rename` underneath, so it shouldn't be used for the same reasons. Its only value-add is calling `mkdirp` implicitly, which is an option in 'mv'.

rightparen let me know if there's any subtleties with the new nuclide-fs stuff here.

Reviewed By: wbinnssmith

Differential Revision: D6025422

fbshipit-source-id: d00676bfaf6a9f98fb14281d0146872408e11706