Add HighlightedText component
Summary:
Extracted from Omni2, this adds a HighlightedText component which
accepts string of text as well as a list of text ranges to be
highlighed.

This also implements an adapter to these lists of ranges from the
slightly awkward FuzzAldrin Plus api of a list of matched characters,
which are often consecutive.

Reviewed By: matthewwithanm

Differential Revision: D6027003

fbshipit-source-id: 603ce7d7ed6c90220df3b55c0e2dce8c4d572ce4