Workaround a bunch of scrolling problems on Windows
Summary:
Scrolling on Windows is messed up, for some reason.
This diff fixes one issue and works around another one.

First, we were setting the `.scrollTop` way too often, on every update of a file tree, rather than just on when it got unhidden (introduced in D5247771)
Another one is weirder. Since the `file-tree` is windowed it rerenders on scrolling, but for some reason on Windows, when scrolling up this causes another `scroll` event to be raised. We end up being in an endless loop constantly scrolling up a bit until we reach the top of the file tree.

Reviewed By: ebluestein

Differential Revision: D6056747

fbshipit-source-id: cf46ddfda66e2b1367dfd3bb98acb36c9bea8a22