Handle tooltip dispose requests that come very quicky after tooltip registration requests
Summary:
If a tooltip disposal requests comes very quickly after the registration request we should remove the registration request. Otherwise, when the `_performMaintenance` callback will be called it wouldn't know which came first and it'll drop both the disposal and the registration.
If, however, the tooltip for that node was already registered this would effectively make the tooltip stuck.

Reviewed By: tjfryan

Differential Revision: D6085502

fbshipit-source-id: 52d5d43a26fbd53caa9e6310d0c8087363d52e49