Wire linterv2 callback-based solutions to diagnostics
Summary:
This adds a new "actions" field inside of `DiagnosticMessage` that allows the passing of LinterV2-style callback solutions.

`FileDiagnosticMessage` cannot contain `actions` due to remotability constraints, so in its place it has a `void` actions field to help Flow with the subtyping.

Due to constraints on how fixes are currently handled, it's still not possible to handle multiple edit-based solutions :(
I think this would require a significant rewrite of how we handle edits in diagnostics today (using the `MessageRangeTracker`, which is kinda ugly).

For now, `actions` will be used purely for callback-based `solutions`.

Diagnostic messages with `actions` can still be queried for additional code actions. I'm not sure if this is a good idea or not, but it seems like the right choice to me.

Reviewed By: pelmers

Differential Revision: D6071516

fbshipit-source-id: 33cd2238fc9fae57489de449e49c9ea488186826