Record processes spawned by Nuclide in its log file.
Summary:
In debugging Nuclide's behavior, particularly on a remote machine, it is helpful
to be able to audit the processes it has spawned. Fortunately, most processes
in Nuclide are spawned via `nuclide-commons/process.js`, so we only need to add
the logging logic in one place.

Incidentally, the `loggedCalls` array where the process info was previously
stored was only read by `nuclide-health`, which only surfaces information about
what happens on the local machine, so it is not sufficient for debugging remote
development. (Historically, this was less of a concern because our primary
concern was ensuring the client side of Atom/Nuclide was responsive.)

Note this adds `log4js` as a dependency of `nuclide-commons`.

Reviewed By: drmathochist

Differential Revision: D6091217

fbshipit-source-id: b5dfb7f240ecc8fb0eb8b310e4568e192f38d183