Start using TextBuffer.onDidChangeText everywhere
Summary:
With multiple cursors, `onDidChange` fires once for each change while `onDidChangeText` aggregates the changes.

It's now recommended to always use this whenever possible, so this modifies all the callsites.

It's actually going to be quite a lot of work to do this for `BufferSubscription` so I filed a TODO there instead.

Reviewed By: matthewwithanm

Differential Revision: D6095400

fbshipit-source-id: ebd44c73c0764e2cbf4cff08b48d3858e708d017