remote server deployment
Summary: Allow big-dig to deploy a server package and manage its installation. It creates a "manifest", essentially a hash of the state of the server installation, to detect corruption of the server and avoid overwriting user data. (The manifest stores files and their stats, including the modified-time and size, but does not hash the contents of the files). A server may be installed to a fresh/empty directory, and a "valid" (satisfying the manifest) installation may be overwritten by a different version. Both are guarded by a user prompt. An installation is "corrupt" if its manifest is damaged/missing or if files have been added, removed, or had their stats changed.

Reviewed By: hansonw

Differential Revision: D5879449

fbshipit-source-id: bddb17d5e4017d8dabc3e3c2e0bd26b94bed4c97