Introduce promptReconnectOnFailure in RemoteConnectionConfiguration
Summary:
Currently, on Nuclide reload, Nuclide attempts to reconnect to the remote connection in the state. If the reconnect fails, Nuclide opens the Remote Connection dialog with the attempted connection in the "Most Recent" tab.
Sometimes, this is undesired. Introducing a property 'promptReconnectOnFailure' in the RemoteConnection config. If promptReconnectOnFailure is false, don't open the connection dialog if attempt connect fails on Nuclide window reload. Otherwise, keep the current behavior if connect attempt fails, prompt the user to connect again with the connection information filled in.

Reviewed By: hansonw

Differential Revision: D6091607

fbshipit-source-id: 5a1ffbabc9813073991517c944fa6917bee8aaa8