Nuclide C++ debugger: only send disassembly if disassembly pane is open
Summary:
When debugging smaller apps, the disassembly view is fast enough to just use all the time. But for larger apps with big frames, the overhead to dumping the frame and sending it over the wire with each frame is too high and is degrading debugging perf.  This change sends the disassembly only if the disassembly pane is actually open. Users can simply close the pane and then pay no cost for it, and open it when it is needed.

If the pane is closed, Nuclide already remembers that and won't open it again unless the user does so explicitly.

Reviewed By: mostafaeweda

Differential Revision: D6079989

fbshipit-source-id: 2878dd646f4d8568703b1b731722ca513613cb41