Make HostServices.showProgress robust against flakey message loss
Summary:
We'd observed that over intermittent connections, showProgress-related messages were being lost by nuclide-rpc.

Typical symptom was that the user would see several orphaned progress notifications in the status bar, and the local logs would show an "out-of-order" message from nuclide-rpc.

I found two causes: (1) client constructs a Progress message, sends it to the server, but the server never receives it and so the Progress indicator was orphaned and never disposed; (2) server requests to dispose the progress message, sends this request to the client, but the client never receives it and again the Progress indicator was orphaned and never disposed.

This diff papers over the issue. It changes it so for every progress creation/disposal, the server will send a message back to the client saying "These are all the outstanding progress messages I expect you to have right now". If the client has any additional ones beyond that list, then their disposals must have gotten lost in the mail, and so the client can freely dispose them.

Reviewed By: hansonw

Differential Revision: D6091910

fbshipit-source-id: 06b23ec38b0a329bfe1465eb87b87b8d7bbc65d2