Cache the results of isNonNfsDirectory() in FuzzyFileSearchService.
Summary:
Looking at the output from D6091217, I noticed there were a number of `stat`
calls as I was typing in the Quick Open box. This was due to this line in
`FuzzyFileSearchService.js`:

```
const isEden = await fsPromise.isNonNfsDirectory(pathToDotEden);
```

This was being called every time I typed a character in Quick Open, adding 2ms
of latency each time according to the log file. (Note that I was developing in
a remote directory, so each `stat` call entailed both an RPC and spawning a new
process. Looking at the implementation of `fsPromise.getFileSystemType()`, we
must use `stat(1)` instead of `stat(2)`.)

In practice, `queryFuzzyFile()` is called with values from
`atom.project.getPaths()`, so it is a very small set over the lifetime of an
instance of Nuclide. This revision introduces a small LRU cache for the
filesystem metadata, which should be adequate. The cache value is set high
enough such that if it is exceeded, something has probably gone wrong. Note that
we log a message in this case.

Reviewed By: hansonw

Differential Revision: D6092926

fbshipit-source-id: 3c3d5730d1023f6b5f64229cf6897e7bf28b4291