Make dropdown options UX more consistent across all combinations
Summary:
Sometimes we would display a single submenu with multiple device groups within. This changes the dropdown to always try to bring nested groups to the top level if there are no other options anyway.

We also start displaying the device group name when a device is selected, which is more correct. Some platform providers could use an update now to tidy up their options, but it makes sense to provides consistent rules across the board.

Reviewed By: ebluestein

Differential Revision: D6136546

fbshipit-source-id: 65bce89e0ad809681462492ed6271fe6058cc675