Prevent hyperclick underlines from being covered by datatip highlights
Summary:
If you cmd-hover a location, the hyperclick underline usually comes in before the datatip (due to the higher datatip debounce).
This leads to the underline being obscured, which can be confusing as it seems like hyperclick is no longer available.

Easy fix: just set a z-index on the hyperclick highlight.

Reviewed By: wbinnssmith

Differential Revision: D6137577

fbshipit-source-id: 55a1991ba9208421adcc3b0dbac29ee6c4d44d6b