Make datatips appear at the triggered location rather than the start of the range
Summary:
This is a common complaint in feedback, especially for diagnostics datatips (which often cover a large range). This is super easy to implement at the cost of an additional marker. I put a diff up since it's an easy fix, but not totally sure if it's the right thing to do :D

Fixes https://github.com/facebook-atom/atom-ide-ui/issues/19.

Reviewed By: jgebhardt

Differential Revision: D6137202

fbshipit-source-id: facb3e13226285af3dd8778315602ad311d4bd75