Allow createMultiLspLanguageService to return null
Summary:
Users of createMultiLspLanguageService pretty much should all check if the command actually exists; otherwise we'll get a bunch of red boxes when we try to spawn it.

This should be handled within `createMultiLspLanguageService` - I've made the return value nullable (along with a warning). Users can create a `NullLanguageService` to create a 'no-op' language service. (I didn't make `createMultiLspLanguageService` return a `NullLanguageService` because then we'd still have to go through a remote object for the "no-op" calls)

Note: I'd like to make `LanguageService` nullable all the way through the stack so we don't even have to construct a `NullLanguageService`. But that's going to be a pretty involved change.

Reviewed By: ljw1004

Differential Revision: D6156813

fbshipit-source-id: 5d23dd7a7418b2faf177257f5a766c40ff50bb30