Improve Table resizing logic
Summary:
At this point, I'm *pretty* confident that there isn't an O(n) way to do this and still satisfy all of our conditions. But there are never that many columns anyway so I think it's okay.

For clarity's sake I also moved away from the imperative management of the resizing events and manual disposable management. I think it's a big improvement!

Plus I wrote a bunch of tests to check the fundamental behaviors we want from this thing. (They're unit tests for the resize algorithms, not integration tests of the component.)

Reviewed By: hansonw

Differential Revision: D6158729

fbshipit-source-id: 8734a05f4da4845e4e7ac5f61299a04be98547a6