Use disposable returned from atom.keymaps.add
Summary:
Rather than manually tracking whether keymaps have been registered in order to not add them multiple times (and failing to unregister them), use and dispose of the disposable returned from this api when the provider is disposed of.

Additionally, update our atom typings to reflect that this returns a disposable.

Reviewed By: matthewwithanm

Differential Revision: D6162589

fbshipit-source-id: 2129cd8da8927be8c111a3426396c533b8742cd2