Create iconForOutlineTree and use it in omni, breadcrumbs
Summary:
This creates a new module in commons-atom called `iconForOutlineTree` which, given an outline tree object, returns the icon name which should be used to render it. This is now used in breadcrumbs.

It also exports a function for returning an icon name given an OutlineTreeKind. This is now used in omni2.

Reviewed By: matthewwithanm

Differential Revision: D6166977

fbshipit-source-id: eaa6fc61d78ef8ea5792e7c52286a25bd1dac2e4