Do not serialize excessive state in file-tree
Summary:
File tree's internal state can grow quite huge. When the serialization was only performed on package deactivation it wasn't much of a problem, but nowdays it's done every few seconds.There's very little benefit in serializing this over and over again.

This removes most of the file-tree's state from being serialized. Writing this data is too expensive and the benefit of having your Nuclide reopened with file-tree already being aware of the files that it had listed for the previous time is not sufficient to justify that huge investment.

After this diff, the file tree will load up with just root directories and will build up its state as it normally should.

Reviewed By: hansonw

Differential Revision: D6130581

fbshipit-source-id: 65a9dabc828b1d686030db7d473ed58808d3e021