Remove forced updates from file-tree rendering logic
Summary:
This removes a forced React update in FileTreeSidebarComponent (`this.forceUpdate()`) and instead utilizes component state to perform the render when necessary.

Previously the rendering logic would peek into the FileTreeStore a lot, post this diff all (well, all but one) instances of `this._store` access were removed from the rendering flow and instead everything is fetched from state. As it should have been done ages ago. The `.forceUpdate` was part of the very-very old design, which predated file-tree rewrite and now I finally got to getting rid of it.

Reviewed By: matthewwithanm

Differential Revision: D6138049

fbshipit-source-id: b1d7c8d24e0445680605e584df40b279d98fb7ec