Auto-fetch and expand root nodes when added, not in FileTreeActions
Summary: We want to auto-expand roots when we add them to the file tree, but we used (for some reason) to do that inside the Actions class. This makes little sense - if we want this behavior we should implement it in the store and keep the Actions free of that nonesense.

Reviewed By: matthewwithanm

Differential Revision: D6164341

fbshipit-source-id: 705713f0b3e2b2ac99ba2505d5bd8f69c52d0ebf