Add timeout to file-tree's node fetching
Summary: When we have a status of loading on any directory in the file tree it affects how we treat it. Most of the time it's a non-issue, but sometimes these calls may get lost and stuck forever. This diff adds a 20sec timeout for this, which should be more than enough for any reasonable fetch.

Reviewed By: wbinnssmith

Differential Revision: D6164370

fbshipit-source-id: 0757539db40084e920a85b4b4055d982d64c754e