Address root cause of HostServices double-dispose
Summary:
In HostServices, the design is that (1) anyone can do forkHostServices to get their own copy to pass onto subsystems, (2) you can dispose of an ancestor HostService, which will dispose of every piece of dialog/progress that it or its children created.

But the subsystems which use a descendent of HostServices currently have no way of knowing that they're using a disposed fork of it!

Should they? Should there be a new API so that each subsystem has to listen for whether its HostServices has been disposed? Or should each individual call that the subsystem makes to its HostServices throw an AlreadyDisposed exception? Or should such calls just be no-ops returning empty results?

I think it's cleanest and easiest if such calls are just no-ops.

That's what I've implemented in this diff. Any operation on a HostServices where it or its parent has been disposed will just be a silent no-op.

Reviewed By: hansonw

Differential Revision: D6139249

fbshipit-source-id: b131e65de4e63abe6c3a2e2a1ace8516842222ee