Only focus the outline view search when the outer visibility changes
Summary:
Handling the visibility as an observable is tricky, since in many cases the change in visibility is what causes the search input to mount.
Using a behavior subject like in the current code works, but it leads to the search box always focusing on mount (e.g. when switching between an editor that doesn't have outline view to one that does). However, with a regular observable subscribing on `componentDidMount` is too late, as the visibility change has already occurred!

Instead, this:

1) switches to `bindObservableAsProps` on the outside to avoid observables altogether
2) uses `componentDidChange` in the outer-most component to detect visibility changes (since it's always mounted)
3) passes focus down with refs.

It's kind of gross to have this chain of refs all the way down to the search box but it's not too much worse than threading an Observable down.

Reviewed By: jgebhardt

Differential Revision: D6202839

fbshipit-source-id: 2e4c30792fb3887a0f555b6c670e65b356ace3ba