Change Breadcrumbs parent indicator icon
Summary:
ebluestein brought up a very good point, which is that the arrow indicator which we are currently using sets up an expectation that hovering over the entry will spawn a nested menu to the right. This is the iconography/behavior on both OSX and Windows.

Since we have decided on not going down the path of nested dropdowns, we should use a different icon which still conveys "this node has children".

I'm opening this diff mostly as a strawman to get ideas for what icon we might use. The goal is to find an icon which:

1. Does _not_ setup the expectation that hovering the entry will expand a nested dropdown.
2. _Does_ convey that this entry has children.

Reviewed By: matthewwithanm

Differential Revision: D6161240

fbshipit-source-id: 6750d9bc3fced891c3d781379104663b2dec1526