Execute terminal commands on selection
Summary:
This executes whatever is entered following the "!" prefix in Omni2.

For now, it always opens them in the bottom dock, as to not obscure the main text area.

It also sets the terminal title to the query that was entered so these are recognizable.

Personally I wonder if these should be implemented as a background process that pipes its output to console instead. cc matthewwithanm

Reviewed By: rightparen

Differential Revision: D6190457

fbshipit-source-id: f19d66c19c8af3f3d4c7c58b7135f2b684eadc98