Move the periodic join functionality to ScribeProcess
Summary:
I thought it was a one-off, but I realized I can use the scribe process for regular analytics as well. This implements the periodic joining inside of the `ScribeProcess` class.

To simplify things a little, I removed `dispose()` (since we never use it!)
`join()` should be used in any case where the Scribe process should be shut down, as it ensure that there's no data loss.

Reviewed By: matthewwithanm

Differential Revision: D6169416

fbshipit-source-id: 8e6be85c5f9eefb7c37d298e7cca3c646a4dfd87