Parse additional data from merge conflicts
Summary:
There are some commands that don't have a easily deducible continue/abort command. For example in the case of rebase, if you encounter merge conflicts then you need to run 'rebase --continue/--abort' but in the case of merge its not just 'merge --continue/--abort'.

This change gets the command to continue or abort from hg itself.

Reviewed By: jgebhardt

Differential Revision: D6199737

fbshipit-source-id: 5ae58b249b33085d67b94010924849952a1274c0