Return a remote directory placeholder when reloading remote projects
Summary:
Starting in Atom 1.22, Atom now checks for `directory.existsSync()` before proceeding with the rest of deserialization (and it throws up a big red error box if any directories no longer exist).

This doesn't actually break any functionality, but it's pretty scary to users. We'll return a dummy `RemoteDirectoryPlaceholder` to work around this.

There's still a bunch of uncaught rejections that happen because Atom attempts to attach file watchers to the remote URI. I'm working on addressing that separately.

Reviewed By: matthewwithanm

Differential Revision: D6215194

fbshipit-source-id: fa725d70d050d989352880a483303edcf5cf9347